requirements.txt
torch==2.2.1
torchvision==0.17.1
numpy>=1.24.3
matplotlib>=3.7.1
tqdm==4.65.0
Pillow>=9.5.0
pandas>=2.0.3
scikit-learn>=1.4.0
tensorboard==2.13.0
torchattacks==3.4.0
foolbox==3.3.3
lpips==0.1.4
opencv-python==4.8.0.76
jupyter==1.0.0
seaborn==0.12.2 